refactor index.js
move drag-drop code into a separate module
